Core #1921 - Remove action menu items based on leos light profile.

See merge request leos/core!1901

+++ modules/js/src/main/js/ui/extension/actionManagerExtension.js
@@ -588,16 +588,16 @@ define(function actionManagerExtensionModule(require) {
 
         let template = ['<div class="leos-actions Vaadin-Icons">']; //FIXME: we can directly create elements
 
-        if (insertBeforeAndAfter) {
+        if (insertBeforeAndAfter && connector.getState().tocEdition) {
             template.push(`<span data-widget-type="insert.before" title="Insert ${type} before">&#xe622</span>`);
         }
         if (editable) {
             template.push(`<span data-widget-type="edit" title="Edit text">&#xe7fa</span>`);
         }
-        if (deletable) {
+        if (deletable && connector.getState().tocEdition) {
             template.push(`<span data-widget-type="delete" title="Delete ${type}">&#xe80b</span>`);
         }
-        if(insertBeforeAndAfter) {
+        if(insertBeforeAndAfter && connector.getState().tocEdition) {
             template.push(`<span style="transform: rotate(180deg);" data-widget-type="insert.after" title="Insert ${type} after">&#xe623</span>`);
         }
